Move getSemesterTimetable to reducer/timetables
