+++ www/src/js/reducers/timetables.js
@@ -1,6 +1,6 @@
 // @flow
 import type { FSA } from 'types/redux';
-import type { ClassNo, LessonType } from 'types/modules';
+import type { ClassNo, LessonType, Semester } from 'types/modules';
 import type { ModuleLessonConfig, SemTimetableConfig } from 'types/timetables';
 import type { ColorMapping, TimetablesState } from 'types/reducers';
 
@@ -10,9 +10,11 @@ import update from 'immutability-helper';
 
 import config from 'config';
 import createPersistConfig from 'storage/createPersistConfig';
-import { ADD_MODULE, REMOVE_MODULE, CHANGE_LESSON, SET_TIMETABLE } from 'actions/timetables';
+import { ADD_MODULE, CHANGE_LESSON, REMOVE_MODULE, SET_TIMETABLE } from 'actions/timetables';
 import { SET_EXPORTED_DATA } from 'actions/export';
 import { getNewColor } from 'utils/colors';
+
+const EMPTY_OBJECT = {};
 
 // Map of LessonType to ClassNo.
 const defaultModuleLessonConfig: ModuleLessonConfig = {};
@@ -42,7 +44,6 @@ function moduleLessonConfig(
 
 // Map of ModuleCode to module lesson config.
 const defaultSemTimetableConfig: SemTimetableConfig = {};
-
 function semTimetable(
   state: SemTimetableConfig = defaultSemTimetableConfig,
   action: FSA,
@@ -68,6 +69,7 @@ function semTimetable(
   }
 }
 
+// Map of semester to color mapping
 const defaultSemColorMap = {};
 function semColors(state: ColorMapping = defaultSemColorMap, action: FSA): ColorMapping {
   const moduleCode = _.get(action, 'payload.moduleCode');
@@ -147,3 +149,14 @@ export function timetablesReducer(
 }
 
 export default persistReducer(createPersistConfig('timetables'), timetablesReducer);
+
+// Extract sem timetable and colors for a specific semester from TimetablesState
+export function getSemesterTimetable(
+  semester: Semester,
+  state: TimetablesState,
+): { timetable: SemTimetableConfig, colors: ColorMapping } {
+  return {
+    timetable: state.timetableConfig[semester] || EMPTY_OBJECT,
+    colors: state.colors[semester] || EMPTY_OBJECT,
+  };
+}
