+++ www/src/js/utils/export.js
@@ -3,7 +3,7 @@ import qs from 'query-string';
 import type { Semester } from 'types/modules';
 import type { State } from 'reducers';
 import type { ExportData } from 'types/export';
-import { getSemesterTimetable } from 'utils/timetables';
+import { getSemesterTimetable } from 'reducers/timetables';
 
 export function extractStateForExport(semester: Semester, state: State): ExportData {
   const { colors, timetable } = getSemesterTimetable(semester, state.timetables);
